// frontend/src/types/city.ts
export interface City {
  id: number;
  city: string;
  country: string;
  continent: string;
  costIndex: number;
  internetSpeed: number;
  safetyScore: number;
  climate: string;
  nomadScore: number;
  rent: number;
  foodCost: number;
  transportCost: number;
  coworkingCost: number;
  temperatureAvg: number;
  latitude: number;
  longitude: number;
  image: string;
}